requirements.txt
ccxt==4.3.89
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
SQLAlchemy>=2.0.0
pyyaml>=6.0
joblib>=1.3.0
python-dateutil>=2.8.2
retrying>=1.3.4
loguru>=0.7.0
requests~=2.31.0
dotenv~=0.9.9
python-dotenv~=1.1.1
catboost>=1.2.5